RSSEC-172 refactory: change file name
