+++ src/core/interfaces/customerResponse.ts
